+++ src/main/java/pl/plajer/villagedefense3/commands/MainCommand.java
@@ -570,7 +570,7 @@ public class MainCommand implements CommandExecutor {
       player.sendMessage("");
       player.sendMessage(ChatColor.GREEN + "Edit this arena via " + ChatColor.GOLD + "/vd " + args[1] + " edit" + ChatColor.GREEN + "!");
       player.sendMessage(ChatColor.GOLD + "Don't know where to start? Check out tutorial video:");
-      player.sendMessage(ChatColor.GOLD + "https://bit.ly/2xwRU8S");
+      player.sendMessage(ChatColor.GOLD + SetupInventory.VIDEO_LINK);
       player.sendMessage(ChatColor.BOLD + "------------------------------------------- ");
     }
   }